+++ packager/react-packager/src/DependencyResolver/FileWatcher/__mocks__/sane.js
@@ -9,5 +9,5 @@
 'use strict';
 
 module.exports = {
-  WatchmanWatcher: jest.genMockFromModule('sane/src/watchman_watcher')
+  WatchmanWatcher: jest.genMockFromModule('sane/src/watchman_watcher'),
 };
