+++ packager/react-packager/src/DependencyResolver/FileWatcher/__tests__/FileWatcher-test.js
@@ -12,28 +12,37 @@ jest
   .dontMock('util')
   .dontMock('events')
   .dontMock('../')
-  .dontMock('q')
   .setMock('child_process', {
     exec: function(cmd, cb) {
       cb(null, '/usr/bin/watchman');
-    }
+    },
   });
 
-var FileWatcher = require('../');
 var sane = require('sane');
 
 describe('FileWatcher', function() {
   var Watcher;
+  var FileWatcher;
+  var config;
 
   beforeEach(function() {
     Watcher = sane.WatchmanWatcher;
     Watcher.prototype.once.mockImplementation(function(type, callback) {
       callback();
     });
+    FileWatcher = require('../');
+
+    config = [{
+      dir: 'rootDir',
+      globs: [
+        '**/*.js',
+        '**/*.json',
+      ],
+    }];
   });
 
   pit('it should get the watcher instance when ready', function() {
-    var fileWatcher = new FileWatcher(['rootDir']);
+    var fileWatcher = new FileWatcher(config);
     return fileWatcher.getWatchers().then(function(watchers) {
       watchers.forEach(function(watcher) {
         expect(watcher instanceof Watcher).toBe(true);
@@ -46,10 +55,10 @@ describe('FileWatcher', function() {
     Watcher.prototype.on.mockImplementation(function(type, callback) {
       cb = callback;
     });
-    var fileWatcher = new FileWatcher(['rootDir']);
+    var fileWatcher = new FileWatcher(config);
     var handler = jest.genMockFn();
     fileWatcher.on('all', handler);
-    return fileWatcher.getWatchers().then(function(){
+    return fileWatcher.getWatchers().then(function() {
       cb(1, 2, 3, 4);
       jest.runAllTimers();
       expect(handler.mock.calls[0]).toEqual([1, 2, 3, 4]);
@@ -57,7 +66,7 @@ describe('FileWatcher', function() {
   });
 
   pit('it should end the watcher', function() {
-    var fileWatcher = new FileWatcher(['rootDir']);
+    var fileWatcher = new FileWatcher(config);
     Watcher.prototype.close.mockImplementation(function(callback) {
       callback();
     });
