+++ packager/react-packager/src/DependencyResolver/FileWatcher/index.js
@@ -12,11 +12,10 @@ const EventEmitter  = require('events').EventEmitter;
 const sane = require('sane');
 const Promise = require('promise');
 const exec = require('child_process').exec;
-const _ = require('underscore');
 
 const MAX_WAIT_TIME = 25000;
 
-// TODO(amasad): can we use watchman version command instead?r
+// TODO(amasad): can we use watchman version command instead?
 const detectingWatcherClass = new Promise(function(resolve) {
   exec('which watchman', function(err, out) {
     if (err || out.length === 0) {
@@ -81,13 +80,10 @@ class FileWatcher extends EventEmitter {
   }
 
   static createDummyWatcher() {
-    const ev = new EventEmitter();
-    _.extend(ev, {
+    return Object.assign(new EventEmitter(), {
       isWatchman: () => Promise.resolve(false),
       end: () => Promise.resolve(),
     });
-
-    return ev;
   }
 }
